% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareGraphs.R
\name{compareGraphs}
\alias{compareGraphs}
\title{Compare the estimated graph with the true graph}
\usage{
compareGraphs(gl, gt)
}
\arguments{
\item{gl}{the estimated graph}

\item{gt}{the true graph}
}
\value{
A vector of three components: true positive rate, false positive 
rate, true discovery rate.
}
\description{
Both graphs should be undirected graphs.
}
