% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_ZINB.R
\name{cond_ZINB}
\alias{cond_ZINB}
\title{Perform the conditional independence test based on the ZINB model}
\usage{
cond_ZINB(x, y, z, alpha)
}
\arguments{
\item{x, y}{variables for testing independence}

\item{z}{conditioning set of variables}

\item{alpha}{the level of the significance for test of independence}
}
\value{
returns the p-value of the test
}
\description{
Check whether x and y are independent conditional on z in the ZINB model.
}
