% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genER.R
\name{genER}
\alias{genER}
\title{Generate data using the Erdos and Renyi (ER) model}
\usage{
genER(p, pE, n, effS = NULL)
}
\arguments{
\item{p}{number of vertices}

\item{pE}{probability that two vertices will connect}

\item{n}{sample size for data generation}

\item{effS}{the effect size, must be a single number - - default is 
\code{NULL} which generates random coefficients for each covariate from a 
Uniform(0,1) distribution.}
}
\value{
A list with the following components:
\describe{
  \item{A}{adjacency matrix of the simulated graph}
  \item{X}{simulated sample data}
  \item{G}{simulated graph as a graphNEL object}
}
}
\description{
Given the number of vertices, the probability that any two vertices connect,
and the sample size, generate the true graph structure and the sample data.
}
\details{
The true graph structure is generated such that the probability of 
having an edge for any pair of vertices is \eqn{p_E = d_0/p}, where \eqn{d_0}
is average number of edges per node. Then, all graphs with p vertices and
d edges have probability of \eqn{p_E^d(1-p_E)^{\choose(d, 2) -d}} to be
generated. After we generate the adjacency matrix \eqn{A}, the sample 
data are simulated from the multivariate normal distribution with
mean 0 and covariance \deqn{ (I - B)^{-1}(I-B^T)^{-1} } where \eqn{B = } the
effect size \eqn{\times} the lower triangular matrix of \eqn{A}.
}
\examples{
genER(10, 0.1, 20, 0.5)
}
