% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marscr.R
\name{marscr}
\alias{marscr}
\title{Perform the marginal screening prior to penalized regression}
\usage{
marscr(data, thres = 0.1)
}
\arguments{
\item{data}{the sample data}

\item{thres}{either a number between 0 and 1 or an positive integer. If it is 
a number between 0 and 1, the function selects the variables whose 
correlation with the response are greater than the number; 
if it is an integer, the function selects the top thres variables. 
Default is 0.1.}
}
\value{
A square matrix. The ith row provides the most correlated variables 
with the ith variable.
}
\description{
This function removes the covariates that are not even weakly associated with 
the response variable prior to penalized regression to improve the performance
of the variable selection and reduce the runtime.
}
\examples{
data <- genER(10, 0.1, 20, 0.5)$X
gEstM <- marscr(data, thres = 5)
}
