% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{mirror}
\alias{mirror}
\title{Create the mirror of a graph.}
\usage{
mirror(graph)
}
\arguments{
\item{graph}{the graph}
}
\value{
returns the mirror of the graph.
}
\description{
Create the mirror of a graph.
}
