% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcZINB.R
\name{pcZINB}
\alias{pcZINB}
\title{Estimate the skeleton of the graphical model using a modified PC algorithm}
\usage{
pcZINB(estG, variables, alpha = 0.01, fun = cond_ZINB)
}
\arguments{
\item{estG}{the estimated moral graph model}

\item{variables}{the sample data}

\item{alpha}{the level of significance used in the conditional independence 
test - default is 0.01}

\item{fun}{the function used for the conditional independence test}
}
\value{
A list consisting of two components:
\describe{
  \item{res}{the removed vertex pairs and their d-separation sets}
  \item{X}{simulated sample data}
  \item{G}{the matrix adjacency matrix for the resulting graph}
}
}
\description{
After estimating moral graph using neighborhood selection, the false 
positive connections can be removed using a modified PC algorithm, which 
removes an edge 
between two variables if they are conditional independent given any subset 
of other variables.
}
