\name{scZINB-package}
\alias{scZINB-package}
\alias{scZINB}
\docType{package}
\title{
\packageTitle{scZINB}
}
\description{
\packageDescription{scZINB}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scZINB}
\packageIndices{scZINB}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{scZINB}

Maintainer: \packageMaintainer{scZINB}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
