% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentp.R
\name{gentp}
\alias{gentp}
\title{Generate tuning parameters for variable selection in the zero-inflated 
negative binomial regression}
\usage{
gentp(nlambda = 10, ntau = 3, y, X, unpenalizedx = NULL,
  unpenalizedz = NULL, offsetx = NULL, offsetz = NULL, pfactor = 0.01,
  penType = 1)
}
\arguments{
\item{nlambda}{number of unique lambda and tau values - default are 10 
and 3.}

\item{ntau}{number of unique lambda and tau values - default are 10 
and 3.}

\item{y}{zero-inflated count response}

\item{X}{covariate matrix. Intercept is added within the function. This 
could take '1' as the input which indicates an intercept-only model.}

\item{unpenalizedx}{Additional unpenalized covariates for 
negative binomial and logistic regression respectively. Default is 
\code{NULL}.}

\item{unpenalizedz}{Additional unpenalized covariates for 
negative binomial and logistic regression respectively. Default is 
\code{NULL}.}

\item{offsetx, offsetz}{Two vector of observations that are included in the 
linear predictors of negative binomial regression and logistic regression 
respectively. Default are \code{NULL}.}

\item{pfactor}{default is 1e-2. The multiplier for the largest calculated 
penalty to determine smallest penalty value. Use in conjunction with
nlambda/ntau to control the granularity of the tp grid.}

\item{penType}{options are 1 (default) or 2. 1 is the group log penalty. 2 
is lasso.}
}
\value{
A matrix with two columns. Each row represents a pair of tuning
parameters: lambda and tau.
}
\description{
Use a dynamic tuning parameter algorithm to generate lambads and taus 
in penalized likelihood estimation of the ZINB model.
}
